django-docker/interview-svc/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic>=2
pydantic-settings>=2
python-jose[cryptography]
supabase>=2

